+++ apps/explorer/src/modules/addresses/components/TableAddressTxRow.vue
@@ -58,7 +58,9 @@
                 >
                   <p :class="[!getType(tx) ? 'success--text mb-0' : 'error--text mb-0']">{{ getSign(tx) }}{{ getShortValue(tx.getValue().toEth()) }}</p>
                   <v-tooltip bottom>
-                    <v-icon slot="activator" small class="info--text text-xs-center ml-1">fa fa-question-circle</v-icon>
+                    <template #activator="data">
+                      <v-icon v-on="data.on" small class="info--text text-xs-center ml-1">fa fa-question-circle</v-icon>
+                    </template>
                     <span>{{
                       formatStr(
                         tx
